# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.UserDTO:
    properties:
      avatar_id:
        type: integer
      created_at:
        type: string
      email:
        type: string
      end_date:
        type: string
      first_analyzer_used_at:
        type: string
      first_log_created_at:
        type: string
      first_name:
        type: string
      goal:
        type: string
      has_created_first_log:
        type: boolean
      has_used_analyzer:
        type: boolean
      id:
        type: integer
      is_admin:
        type: boolean
      last_name:
        type: string
      start_date:
        type: string
    type: object
...
  handlers.credentials:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.dashboardResponse:
    properties:
      average_month_karma:
        type: number
      current_streak_days:
        type: integer
      day_karma:
        type: number
      entries_this_week:
        type: integer
      entries_this_year:
        type: integer
      has_today_entry:
        type: boolean
      last7_days_trend:
        items:
          $ref: '#/definitions/handlers.trendPoint'
        type: array
      month_karma:
        type: number
      reference_date:
        type: string
      user:
        $ref: '#/definitions/handlers.UserDTO'
      week_karma:
        type: number
      year_karma:
        type: number
    type: object
...
  handlers.submissionHistoryPoint:
    properties:
      has_submission:
        type: boolean
      local_date:
        type: string
    type: object
  handlers.submissionHistoryResponse:
    properties:
      end_date:
        type: string
      history:
        items:
          $ref: '#/definitions/handlers.submissionHistoryPoint'
        type: array
      start_date:
        type: string
    type: object
  handlers.trendPoint:
    properties:
      karma:
        type: number
      local_date:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for journaling, goal tracking, and personal development metrics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Winning Son API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.credentials'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: User login
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.credentials'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: User signup
      tags:
      - auth
  /dashboard:
    get:
      consumes:
      - application/json
      description: Aggregates and useful metrics to power the dashboard
      parameters:
      - description: Reference date (YYYY-MM-DD)
        in: query
        name: local_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.dashboardResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get dashboard metrics
      tags:
      - dashboard
  /dashboard/submission-history:
    get:
      consumes:
      - application/json
      description: Returns whether the user has submissions for each day in a date
        range (max 365 days)
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.submissionHistoryResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get submission history
      tags:
      - dashboard
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"